refactor(server): split issue handler god-file and fix handler/service boundary (WOR-63)
W2 — file split:
- issue.go trimmed from 3,139 → 176 lines (DTOs + mappers only)
- 8 new sibling files in package handler: issue_search, issue_list,
  issue_grouped, issue_children, issue_quick_create, issue_create,
  issue_update, issue_batch

W1 — handler/service boundary fix:
- Added IssueService.Update + ReconcileAfterUpdate + types
  (IssueUpdateOpts, IssueChangeSummary, IssueUpdateResult) + errors
  (ErrSelfParent, ErrCircularParent) to service/issue.go
- UpdateIssue and BatchUpdateIssues now delegate through the service:
  Update → handler publish → ReconcileAfterUpdate → notifyParentOfChildDone
- Ordering preserved exactly: DB write → attachment link → publish →
  reconcile (cancel/enqueue/promote) → notify
- Batch payload divergence preserved (3-flag subset for batch)
- Removed dead squad predicate trio from handler/squad.go
  (shouldEnqueueSquadLeaderOnAssign, isSquadLeaderReady,
  enqueueSquadLeaderTask) — service has its own copies
- Agent predicate copies (shouldEnqueueAgentTask, isAgentAssigneeReady)
  retained — onboarding_shim.go still calls them

W3 — SQL builder extraction:
- Extracted issueQueryBuilder type with addArg method, de-duplicating
  the closure pattern in ListIssues and ListGroupedIssues

Tests:
- TestReconcileAfterUpdate_AssignBranchCancelsAndEnqueues
- TestReconcileAfterUpdate_BacklogPromotionEnqueues
- TestReconcileAfterUpdate_CancelBranchCancelsTasks
- TestComputeIssueChanges_* (6 sub-tests for change detection)

Verification:
- go build ./... clean
- go vet ./internal/handler/ ./internal/service/ clean
- All new + existing batch/reconcile tests pass
