+++ server/internal/handler/issue_search.go
@@ -1,0 +1,514 @@
+package handler
+
+import (
+	"fmt"
+	"log/slog"
+	"net/http"
+	"regexp"
+	"strconv"
+	"strings"
+	"unicode"
+
+	"github.com/jackc/pgx/v5/pgtype"
+	db "github.com/multica-ai/multica/server/pkg/db/generated"
+)
+
+type SearchIssueResponse struct {
+	IssueResponse
+	MatchSource               string  `json:"match_source"`
+	MatchedSnippet            *string `json:"matched_snippet,omitempty"`
+	MatchedDescriptionSnippet *string `json:"matched_description_snippet,omitempty"`
+	MatchedCommentSnippet     *string `json:"matched_comment_snippet,omitempty"`
+}
+
+// extractSnippet extracts a snippet of text around the first occurrence of query.
+// Returns up to ~120 runes centered on the match. Uses rune-based slicing to
+// avoid splitting multi-byte UTF-8 characters (important for CJK content).
+// For multi-word queries, tries phrase match first; if not found, locates the
+// earliest occurring individual term and centers the snippet around it.
+func extractSnippet(content, query string) string {
+	runes := []rune(content)
+	lowerRunes := []rune(strings.ToLower(content))
+	queryRunes := []rune(strings.ToLower(query))
+
+	idx := findRuneSubstring(lowerRunes, queryRunes)
+
+	// If phrase not found, try individual terms for multi-word queries.
+	matchLen := len(queryRunes)
+	if idx < 0 {
+		terms := strings.Fields(strings.ToLower(query))
+		if len(terms) > 1 {
+			earliest := -1
+			earliestLen := 0
+			for _, term := range terms {
+				termRunes := []rune(term)
+				pos := findRuneSubstring(lowerRunes, termRunes)
+				if pos >= 0 && (earliest < 0 || pos < earliest) {
+					earliest = pos
+					earliestLen = len(termRunes)
+				}
+			}
+			if earliest >= 0 {
+				idx = earliest
+				matchLen = earliestLen
+			}
+		}
+	}
+
+	if idx < 0 {
+		if len(runes) > 120 {
+			return string(runes[:120]) + "..."
+		}
+		return content
+	}
+	start := idx - 40
+	if start < 0 {
+		start = 0
+	}
+	end := idx + matchLen + 80
+	if end > len(runes) {
+		end = len(runes)
+	}
+	snippet := string(runes[start:end])
+	if start > 0 {
+		snippet = "..." + snippet
+	}
+	if end < len(runes) {
+		snippet = snippet + "..."
+	}
+	return snippet
+}
+
+// findRuneSubstring returns the index of needle in haystack, or -1 if not found.
+func findRuneSubstring(haystack, needle []rune) int {
+	if len(needle) == 0 || len(haystack) < len(needle) {
+		return -1
+	}
+	for i := 0; i <= len(haystack)-len(needle); i++ {
+		match := true
+		for j := range needle {
+			if haystack[i+j] != needle[j] {
+				match = false
+				break
+			}
+		}
+		if match {
+			return i
+		}
+	}
+	return -1
+}
+
+// descriptionContains checks if the description text contains the search phrase or all terms.
+func descriptionContains(desc pgtype.Text, phrase string, terms []string) bool {
+	if !desc.Valid || desc.String == "" {
+		return false
+	}
+	lower := strings.ToLower(desc.String)
+	if strings.Contains(lower, strings.ToLower(phrase)) {
+		return true
+	}
+	if len(terms) > 1 {
+		for _, t := range terms {
+			if !strings.Contains(lower, strings.ToLower(t)) {
+				return false
+			}
+		}
+		return true
+	}
+	return false
+}
+
+// escapeLike escapes LIKE special characters (%, _, \) in user input.
+func escapeLike(s string) string {
+	s = strings.ReplaceAll(s, `\`, `\\`)
+	s = strings.ReplaceAll(s, `%`, `\%`)
+	s = strings.ReplaceAll(s, `_`, `\_`)
+	return s
+}
+
+// splitSearchTerms splits a query into individual search terms, filtering empty strings.
+func splitSearchTerms(q string) []string {
+	fields := strings.FieldsFunc(q, func(r rune) bool {
+		return unicode.IsSpace(r)
+	})
+	terms := make([]string, 0, len(fields))
+	for _, f := range fields {
+		if f != "" {
+			terms = append(terms, f)
+		}
+	}
+	return terms
+}
+
+// identifierNumberRe matches patterns like "MUL-123" or "ABC-45".
+var identifierNumberRe = regexp.MustCompile(`(?i)^[a-z]+-(\d+)$`)
+
+// parseQueryNumber extracts an issue number from the query if it looks like
+// an identifier (e.g. "MUL-123") or a bare number (e.g. "123").
+func parseQueryNumber(q string) (int, bool) {
+	q = strings.TrimSpace(q)
+	// Check for identifier pattern like "MUL-123"
+	if m := identifierNumberRe.FindStringSubmatch(q); m != nil {
+		if n, err := strconv.Atoi(m[1]); err == nil && n > 0 {
+			return n, true
+		}
+	}
+	// Check for bare number
+	if n, err := strconv.Atoi(q); err == nil && n > 0 {
+		return n, true
+	}
+	return 0, false
+}
+
+// searchResult holds a raw row from the dynamic search query.
+type searchResult struct {
+	issue                 db.Issue
+	totalCount            int64
+	matchSource           string
+	matchedCommentContent string
+}
+
+// buildSearchQuery builds a dynamic SQL query for issue search.
+// It uses LOWER(column) LIKE for case-insensitive matching compatible with pg_bigm 1.2 GIN indexes.
+// Search patterns are lowercased in Go to avoid redundant LOWER() on the pattern side in SQL.
+// LIKE patterns are pre-built in Go (e.g. "%html%") so pg_bigm can extract bigrams from a single parameter value.
+func buildSearchQuery(phrase string, terms []string, queryNum int, hasNum bool, includeClosed bool) (string, []any) {
+	// Lowercase in Go so SQL only needs LOWER() on the column side.
+	phrase = strings.ToLower(phrase)
+	for i, t := range terms {
+		terms[i] = strings.ToLower(t)
+	}
+
+	// Parameter index tracker
+	argIdx := 1
+	args := []any{}
+	nextArg := func(val any) string {
+		args = append(args, val)
+		s := fmt.Sprintf("$%d", argIdx)
+		argIdx++
+		return s
+	}
+
+	escapedPhrase := escapeLike(phrase)
+	// $1: exact phrase (for exact title match)
+	phraseParam := nextArg(escapedPhrase)
+	// $2: "%phrase%" (contains pattern — pre-built for pg_bigm index usage)
+	phraseContainsParam := nextArg("%" + escapedPhrase + "%")
+	// $3: "phrase%" (starts-with pattern)
+	phraseStartsWithParam := nextArg(escapedPhrase + "%")
+
+	wsParam := nextArg(nil) // $4 — workspace_id, will be filled by caller position
+
+	// Build per-term LIKE conditions only for multi-word search.
+	var termContainsParams []string
+	if len(terms) > 1 {
+		for _, t := range terms {
+			et := escapeLike(t)
+			termContainsParams = append(termContainsParams, nextArg("%"+et+"%"))
+		}
+	}
+
+	// --- WHERE clause ---
+	var whereParts []string
+
+	// Full phrase match: title, description, or comment
+	phraseMatch := fmt.Sprintf(
+		"(LOWER(i.title) LIKE %s OR LOWER(COALESCE(i.description, '')) LIKE %s OR EXISTS (SELECT 1 FROM comment c WHERE c.issue_id = i.id AND LOWER(c.content) LIKE %s))",
+		phraseContainsParam, phraseContainsParam, phraseContainsParam,
+	)
+	whereParts = append(whereParts, phraseMatch)
+
+	// Multi-word AND match (each term must appear somewhere)
+	if len(termContainsParams) > 1 {
+		var termConditions []string
+		for _, tp := range termContainsParams {
+			termConditions = append(termConditions, fmt.Sprintf(
+				"(LOWER(i.title) LIKE %s OR LOWER(COALESCE(i.description, '')) LIKE %s OR EXISTS (SELECT 1 FROM comment c WHERE c.issue_id = i.id AND LOWER(c.content) LIKE %s))",
+				tp, tp, tp,
+			))
+		}
+		whereParts = append(whereParts, "("+strings.Join(termConditions, " AND ")+")")
+	}
+
+	// Number match
+	numParam := ""
+	if hasNum {
+		numParam = nextArg(queryNum)
+		whereParts = append(whereParts, fmt.Sprintf("i.number = %s", numParam))
+	}
+
+	whereClause := "(" + strings.Join(whereParts, " OR ") + ")"
+
+	if !includeClosed {
+		whereClause += " AND i.status NOT IN ('done', 'cancelled')"
+	}
+
+	// --- ORDER BY clause ---
+	// Build ranking CASE with fine-grained tiers.
+	var rankCases []string
+
+	// Tier 0: Identifier exact match
+	if hasNum {
+		rankCases = append(rankCases, fmt.Sprintf("WHEN i.number = %s THEN 0", numParam))
+	}
+
+	// Tier 1: Exact title match
+	rankCases = append(rankCases, fmt.Sprintf("WHEN LOWER(i.title) = %s THEN 1", phraseParam))
+
+	// Tier 2: Title starts with phrase
+	rankCases = append(rankCases, fmt.Sprintf("WHEN LOWER(i.title) LIKE %s THEN 2", phraseStartsWithParam))
+
+	// Tier 3: Title contains phrase
+	rankCases = append(rankCases, fmt.Sprintf("WHEN LOWER(i.title) LIKE %s THEN 3", phraseContainsParam))
+
+	// Tier 4: Title matches all words (multi-word only)
+	if len(termContainsParams) > 1 {
+		var titleTerms []string
+		for _, tp := range termContainsParams {
+			titleTerms = append(titleTerms, fmt.Sprintf("LOWER(i.title) LIKE %s", tp))
+		}
+		rankCases = append(rankCases, fmt.Sprintf("WHEN (%s) THEN 4", strings.Join(titleTerms, " AND ")))
+	}
+
+	// Tier 5: Description contains phrase
+	rankCases = append(rankCases, fmt.Sprintf("WHEN LOWER(COALESCE(i.description, '')) LIKE %s THEN 5", phraseContainsParam))
+
+	// Tier 6: Description matches all words (multi-word only)
+	if len(termContainsParams) > 1 {
+		var descTerms []string
+		for _, tp := range termContainsParams {
+			descTerms = append(descTerms, fmt.Sprintf("LOWER(COALESCE(i.description, '')) LIKE %s", tp))
+		}
+		rankCases = append(rankCases, fmt.Sprintf("WHEN (%s) THEN 6", strings.Join(descTerms, " AND ")))
+	}
+
+	// Tier 7: Comment contains phrase
+	rankCases = append(rankCases, fmt.Sprintf("WHEN EXISTS (SELECT 1 FROM comment c WHERE c.issue_id = i.id AND LOWER(c.content) LIKE %s) THEN 7", phraseContainsParam))
+
+	// Tier 8: Comment matches all words (multi-word only)
+	if len(termContainsParams) > 1 {
+		var commentTerms []string
+		for _, tp := range termContainsParams {
+			commentTerms = append(commentTerms, fmt.Sprintf("LOWER(c.content) LIKE %s", tp))
+		}
+		rankCases = append(rankCases, fmt.Sprintf("WHEN EXISTS (SELECT 1 FROM comment c WHERE c.issue_id = i.id AND (%s)) THEN 8", strings.Join(commentTerms, " AND ")))
+	}
+
+	rankExpr := "CASE " + strings.Join(rankCases, " ") + " ELSE 9 END"
+
+	// Status priority: active issues first
+	statusRank := `CASE i.status
+		WHEN 'in_progress' THEN 0
+		WHEN 'in_review' THEN 1
+		WHEN 'todo' THEN 2
+		WHEN 'blocked' THEN 3
+		WHEN 'backlog' THEN 4
+		WHEN 'done' THEN 5
+		WHEN 'cancelled' THEN 6
+		ELSE 7
+	END`
+
+	// --- match_source expression ---
+	matchSourceExpr := fmt.Sprintf(`CASE
+		WHEN LOWER(i.title) LIKE %s THEN 'title'
+		WHEN LOWER(COALESCE(i.description, '')) LIKE %s THEN 'description'
+		ELSE 'comment'
+	END`, phraseContainsParam, phraseContainsParam)
+
+	// For multi-word: also check if all terms match in title/description
+	if len(termContainsParams) > 1 {
+		var titleTerms []string
+		var descTerms []string
+		for _, tp := range termContainsParams {
+			titleTerms = append(titleTerms, fmt.Sprintf("LOWER(i.title) LIKE %s", tp))
+			descTerms = append(descTerms, fmt.Sprintf("LOWER(COALESCE(i.description, '')) LIKE %s", tp))
+		}
+		matchSourceExpr = fmt.Sprintf(`CASE
+			WHEN LOWER(i.title) LIKE %s THEN 'title'
+			WHEN (%s) THEN 'title'
+			WHEN LOWER(COALESCE(i.description, '')) LIKE %s THEN 'description'
+			WHEN (%s) THEN 'description'
+			ELSE 'comment'
+		END`,
+			phraseContainsParam, strings.Join(titleTerms, " AND "),
+			phraseContainsParam, strings.Join(descTerms, " AND "),
+		)
+	}
+
+	// --- matched_comment_content subquery ---
+	// Always return matching comment content regardless of match_source,
+	// so frontend can display comment snippet alongside title/description matches.
+	commentSubquery := fmt.Sprintf(`COALESCE(
+		(SELECT c.content FROM comment c
+		 WHERE c.issue_id = i.id AND LOWER(c.content) LIKE %s
+		 ORDER BY c.created_at DESC LIMIT 1),
+		''
+	)`, phraseContainsParam)
+
+	if len(termContainsParams) > 1 {
+		var commentTerms []string
+		for _, tp := range termContainsParams {
+			commentTerms = append(commentTerms, fmt.Sprintf("LOWER(c.content) LIKE %s", tp))
+		}
+		commentSubquery = fmt.Sprintf(`COALESCE(
+			(SELECT c.content FROM comment c
+			 WHERE c.issue_id = i.id AND (LOWER(c.content) LIKE %s OR (%s))
+			 ORDER BY c.created_at DESC LIMIT 1),
+			''
+		)`, phraseContainsParam, strings.Join(commentTerms, " AND "))
+	}
+
+	limitParam := nextArg(nil)  // placeholder
+	offsetParam := nextArg(nil) // placeholder
+
+	query := fmt.Sprintf(`SELECT i.id, i.workspace_id, i.title, i.description, i.status, i.priority,
+		i.assignee_type, i.assignee_id, i.creator_type, i.creator_id,
+		i.parent_issue_id, i.acceptance_criteria, i.context_refs, i.position,
+		i.start_date, i.due_date, i.created_at, i.updated_at, i.number, i.project_id,
+		COUNT(*) OVER() AS total_count,
+		%s AS match_source,
+		%s AS matched_comment_content
+	FROM issue i
+	WHERE i.workspace_id = %s AND %s
+	ORDER BY %s, %s, i.updated_at DESC
+	LIMIT %s OFFSET %s`,
+		matchSourceExpr,
+		commentSubquery,
+		wsParam,
+		whereClause,
+		rankExpr,
+		statusRank,
+		limitParam,
+		offsetParam,
+	)
+
+	return query, args
+}
+
+func (h *Handler) SearchIssues(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	workspaceID := h.resolveWorkspaceID(r)
+
+	q := r.URL.Query().Get("q")
+	if q == "" {
+		writeError(w, http.StatusBadRequest, "q parameter is required")
+		return
+	}
+
+	limit := 20
+	offset := 0
+	if l := r.URL.Query().Get("limit"); l != "" {
+		if v, err := strconv.Atoi(l); err == nil && v > 0 {
+			limit = v
+		}
+	}
+	if limit > 50 {
+		limit = 50
+	}
+	if o := r.URL.Query().Get("offset"); o != "" {
+		if v, err := strconv.Atoi(o); err == nil && v >= 0 {
+			offset = v
+		}
+	}
+
+	includeClosed := r.URL.Query().Get("include_closed") == "true"
+
+	wsUUID, ok := parseUUIDOrBadRequest(w, workspaceID, "workspace_id")
+	if !ok {
+		return
+	}
+	terms := splitSearchTerms(q)
+	queryNum, hasNum := parseQueryNumber(q)
+
+	sqlQuery, args := buildSearchQuery(q, terms, queryNum, hasNum, includeClosed)
+	// Fill placeholder args: $4 = workspace_id, last two = limit, offset
+	args[3] = wsUUID
+	args[len(args)-2] = limit
+	args[len(args)-1] = offset
+
+	rows, err := h.DB.Query(ctx, sqlQuery, args...)
+	if err != nil {
+		slog.Warn("search issues failed", "error", err, "workspace_id", workspaceID, "query", q)
+		writeError(w, http.StatusInternalServerError, "failed to search issues")
+		return
+	}
+	defer rows.Close()
+
+	var results []searchResult
+	for rows.Next() {
+		var sr searchResult
+		if err := rows.Scan(
+			&sr.issue.ID,
+			&sr.issue.WorkspaceID,
+			&sr.issue.Title,
+			&sr.issue.Description,
+			&sr.issue.Status,
+			&sr.issue.Priority,
+			&sr.issue.AssigneeType,
+			&sr.issue.AssigneeID,
+			&sr.issue.CreatorType,
+			&sr.issue.CreatorID,
+			&sr.issue.ParentIssueID,
+			&sr.issue.AcceptanceCriteria,
+			&sr.issue.ContextRefs,
+			&sr.issue.Position,
+			&sr.issue.StartDate,
+			&sr.issue.DueDate,
+			&sr.issue.CreatedAt,
+			&sr.issue.UpdatedAt,
+			&sr.issue.Number,
+			&sr.issue.ProjectID,
+			&sr.totalCount,
+			&sr.matchSource,
+			&sr.matchedCommentContent,
+		); err != nil {
+			slog.Warn("search issues scan failed", "error", err)
+			writeError(w, http.StatusInternalServerError, "failed to search issues")
+			return
+		}
+		results = append(results, sr)
+	}
+	if err := rows.Err(); err != nil {
+		slog.Warn("search issues rows error", "error", err)
+		writeError(w, http.StatusInternalServerError, "failed to search issues")
+		return
+	}
+
+	var total int64
+	if len(results) > 0 {
+		total = results[0].totalCount
+	}
+
+	prefix := h.getIssuePrefix(ctx, wsUUID)
+	resp := make([]SearchIssueResponse, len(results))
+	for i, sr := range results {
+		sir := SearchIssueResponse{
+			IssueResponse: issueToResponse(sr.issue, prefix),
+			MatchSource:   sr.matchSource,
+		}
+		// Always populate comment snippet when a matching comment exists
+		if sr.matchedCommentContent != "" {
+			snippet := extractSnippet(sr.matchedCommentContent, q)
+			sir.MatchedCommentSnippet = &snippet
+			// Keep backward compat: also set MatchedSnippet for comment-source matches
+			if sr.matchSource == "comment" {
+				sir.MatchedSnippet = &snippet
+			}
+		}
+		// Populate description snippet when description matches
+		if sr.matchSource == "description" || descriptionContains(sr.issue.Description, q, terms) {
+			if sr.issue.Description.Valid && sr.issue.Description.String != "" {
+				snippet := extractSnippet(sr.issue.Description.String, q)
+				sir.MatchedDescriptionSnippet = &snippet
+			}
+		}
+		resp[i] = sir
+	}
+
+	w.Header().Set("X-Total-Count", strconv.FormatInt(total, 10))
+	writeJSON(w, http.StatusOK, map[string]any{
+		"issues": resp,
+		"total":  total,
+	})
+}
